# grapevine-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "grapevine API",
    "description": "Decentralized Data Feeds Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "List categories",
        "description": "Retrieve a cursor-paginated list of all categories with optional filtering by active status and search term. Results are ordered by name ASC (alphabetically). Categories are pre-defined in the system and cannot be created via API.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "default": "20"
            },
            "required": false,
            "name": "page_size",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page_token",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "is_active",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories with cursor pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "created_at": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true
                          },
                          "updated_at": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "description": {
                            "type": "string",
                            "nullable": true,
                            "maxLength": 1000
                          },
                          "icon_url": {
                            "type": "string",
                            "nullable": true,
                            "maxLength": 2048,
                            "format": "uri"
                          },
                          "is_active": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "created_at",
                          "updated_at",
                          "id",
                          "name",
                          "description",
                          "icon_url",
                          "is_active"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page_size": {
                          "type": "integer"
                        },
                        "next_page_token": {
                          "type": "string",
                          "nullable": true
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page_size",
                        "next_page_token",
                        "has_more"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/feeds": {
      "get": {
        "tags": [
          "Feeds"
        ],
        "summary": "List feeds",
        "description": "Retrieve a cursor-paginated list of data feeds with optional filtering by owner, category, tags, entry count, age, and active status. Defaults to showing only active feeds. Results are ordered by created_at DESC (newest first).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "default": "20"
            },
            "required": false,
            "name": "page_size",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page_token",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "owner_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "tags",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "default": "1"
            },
            "required": false,
            "name": "min_entries",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "required": false,
            "name": "min_age",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            },
            "required": false,
            "name": "max_age",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "is_active",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of feeds with cursor pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "owner_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "category_id": {
                            "type": "string",
                            "nullable": true,
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "description": {
                            "type": "string",
                            "nullable": true,
                            "maxLength": 2000
                          },
                          "image_cid": {
                            "type": "string",
                            "nullable": true,
                            "maxLength": 255
                          },
                          "is_active": {
                            "type": "boolean"
                          },
                          "total_entries": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "total_purchases": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "total_revenue": {
                            "type": "string",
                            "maxLength": 78
                          },
                          "tags": {
                            "type": "array",
                            "nullable": true,
                            "items": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 50
                            },
                            "maxItems": 20
                          },
                          "created_at": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true
                          },
                          "updated_at": {
                            "type": "integer",
                            "minimum": 0,
                            "exclusiveMinimum": true
                          },
                          "owner_wallet_address": {
                            "type": "string",
                            "minLength": 42,
                            "maxLength": 42,
                            "pattern": "^0x[0-9a-fA-F]{40}$"
                          }
                        },
                        "required": [
                          "id",
                          "owner_id",
                          "category_id",
                          "name",
                          "description",
                          "image_cid",
                          "is_active",
                          "total_entries",
                          "total_purchases",
                          "total_revenue",
                          "tags",
                          "created_at",
                          "updated_at",
                          "owner_wallet_address"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page_size": {
                          "type": "integer"
                        },
                        "next_page_token": {
                          "type": "string",
                          "nullable": true
                        },
                        "has_more": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page_size",
                        "next_page_token",
                        "has_more"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Feeds"
        ],
        "summary": "Create feed",
        "description": "Create a new data feed with specified category, name, description, tags, and optional image. Creates a Pinata storage group using the group ID as the feed ID. The authenticated wallet becomes the feed owner and is automatically created on Base network if it does not exist. Existing wallets must be on Base chain. Enforces maximum feed limit per wallet. Requires wallet authentication via x-payment headers.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "X402 Payment header for payment verification",
              "example": "eyJ2ZXJzaW9uIjoxLCJuZXR3b3JrIjoiYmFzZSIsInBheWxvYWQiOnsicHJpY2UiOnsicGF5bWVudF9pbnN0cnVjdGlvbjoiMGU2YzI3YjAtY2E3ZS00Y2Y0LTg4ZjItY2E0ZjI3ZjYzNzA0IiwiYW1vdW50IjoiMTAwMDAwMDAwMDAiLCJhc3NldCI6IndlYiIsInBheV90byI6IjB4NzQyZDM1Q2M2NjRDMDA1MzI5MjVhM2I4NDRiYmM5ZTU5NWYwYkViIiwiYXV0aG9yaXphdGlvbiI6eyJ0eXBlIjoiRXZtQXV0aCIsImZyb20iOiIweDg2RjE2QjEwNDhDMjRCMkE3NDU2RkE3QjA3RkE3N0E3IiwidG8iOiIweDc0MmQzNUNjNjY0QzA1MzI5MjVhM2I4NDRiYjljZTdlNTk1ZjBiZWIiLCibm9uY2UiOiIxMjM0NTY3ODkwIn19fX0="
            },
            "required": true,
            "description": "X402 Payment header for payment verification",
            "name": "x-payment",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "image_url": {
                    "type": "string",
                    "maxLength": 52428800
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    },
                    "maxItems": 20
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feed created with database-generated UUID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "owner_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "category_id": {
                      "type": "string",
                      "nullable": true,
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "description": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 2000
                    },
                    "image_cid": {
                      "type": "string",
                      "nullable": true,
                      "maxLength": 255
                    },
                    "is_active": {
                      "type": "boolean"
                    },
                    "total_entries": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "total_purchases": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "total_revenue": {
                      "type": "string",
                      "maxLength": 78
                    },
                    "tags": {
                      "type": "array",
                      "nullable": true,
                      "items": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50
                      },
                      "maxItems": 20
                    },
                    "created_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "updated_at": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "owner_wallet_address": {
                      "type": "string",
                      "minLength": 42,
                      "maxLength": 42,
                      "pattern": "^0x[0-9a-fA-F]{40}$"
                    }
                  },
                  "required": [
                    "id",
                    "owner_id",
                    "category_id",
                    "name",
                    "description",
                    "image_cid",
                    "is_active",
                    "total_entries",
                    "total_purchases",
                    "total_revenue",
                    "tags",
                    "created_at",
                    "updated_at",
                    "owner_wallet_address"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - signature verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - maximum feed limit reached or invalid wallet network",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable - failed to create storage group",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    },
                    "details": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}